Update config.yml (#3)
Send users to the `/new/choose` page of discussions when they click on "Questions" or "Features". The idea is that a user already knows they want to ask a question or feature so this is one less click for them. They don't need to land on the generic discussions page and then realize to click "new discussion".