PR template: move description section to top, remove heading
Problem: When I have finished a patch set and want to create a PR,
my desire is to open the link and start typing my reasoning immediately.
However, I find the prompt to fill out entries for changelog and docs
first distracting.

Solution: remove the "Description" heading and move the description
section to the top. When the page is opened, the cursor appears in the
description area, and it's seamless to start typing.

Also, if a PR consists of one commit with a message, that message
will automatically appear where it should in the PR description.